requirements.txt
crewai>=0.126.0; python_version<'3.14'
openai>=1.30.0
fastapi>=0.111.0
uvicorn>=0.30.0
gradio>=5.0.0
pdfplumber>=0.11.0
python-docx>=1.1.2
pydantic>=2.7.0
python-multipart>=0.0.9
PyYAML>=6.0.1
